Merge Indexes (#31)
I completed functionality to merge the different index.rst which exist for the LFOM and entrance tank. The same functionality was introduced for Treatment_Process.rst.

This is accomplished by comparing the two files, index.rst and new_index.rst section by section and adding pieces which exist in index.rst but are missing from new_index.rst . At the end, the one which was added to is maintained as index.rst and new_index.rst is deleted.

In the future the efficiency of the algorithm can definitely be improved. Doctests are included but more complex tests are required as well to ensure edge cases aren't missing.